Fix MSVC compiler warnings
reg. 'argument': conversion from 'size_t' to 'uint32_t/int', possible loss of data (C4267)

+++ src/framework/audio/internal/encoders/flacencoder.cpp
@@ -97,7 +97,7 @@ size_t FlacEncoder::encode(samples_t samplesPerChannel, const float* input)
 
     size_t result = 0;
     size_t totalSamplesNumber = samplesPerChannel * m_format.audioChannelsNumber;
-    size_t frameSize = 1024;
+    uint32_t frameSize = 1024;
     size_t stepSize = frameSize * m_format.audioChannelsNumber;
 
     std::vector<FLAC__int32> buff(samplesPerChannel * sizeof(float));
